requirements.txt
streamlit
google-generativeai
boto3
pathlib 
